+++ net.projecteuler.solutions/src/main/java/utils/Factors.java
@@ -4,10 +4,10 @@ import java.util.Collections;
 import java.util.Set;
 
 public class Factors {
-    public static Set<Integer> of(int n) {
+    public static Set<Long> of(long n) {
         if (n < 1) return Collections.emptySet();
 
-        Set<Integer> s = ProperDivisors.of(n);
+        Set<Long> s = ProperDivisors.of(n);
         s.add(n);
 
         return s;
